Better fix for blinking badge
