+++ src/scripts/browser/managers/notif-manager.js
@@ -4,7 +4,7 @@ class NotifManager extends EventEmitter {
 
   constructor () {
     super();
-    this.unreadCount = 0;
+    this.unreadCount = ''; // number as string; empty string means 0
   }
 
 }
